+++ crates/rustc_codegen_gpu/tests/codegen_fails/implicit_diversed_shared.rs
@@ -7,7 +7,7 @@
 
 #[gpu::kernel]
 #[no_mangle]
-pub fn alloc_shared(a: &[u8], _a_window: usize, b: &mut [u8], b_window: u32, f: &mut [f32], salloc: gpu::DynamicSharedAlloc) {
+pub fn alloc_shared(a: &[u8], _a_window: usize, b: &mut [u8], b_window: usize, f: &mut [f32], salloc: gpu::DynamicSharedAlloc) {
     let mut salloc = salloc;
     let mut dy_shared = salloc.alloc::<f32>(gpu::thread_id::<gpu::DimX>() as usize); //~ ERROR Invalid use of diversed data in GPU code
     //~| ERROR Invalid use of diversed data in GPU code
